+++ src/main/java/org/eclipse/keypop/storagecard/SCInvalidCardResponseException.java
@@ -11,42 +11,39 @@
  ************************************************************************************** */
 package org.eclipse.keypop.storagecard;
 
-import org.eclipse.keypop.reader.CardCommunicationException;
+import org.eclipse.keypop.reader.selection.InvalidCardResponseException;
 import org.eclipse.keypop.storagecard.card.StorageCard;
 
 /**
- * Indicates an input/output error that occurred while communicating with the {@link StorageCard}.
+ * Indicates that a command returned an unexpected or invalid status while interacting with the
+ * {@link StorageCard}.
  *
- * <p>This exception reflects errors in the card communication process, such as transmission
- * failures, card removal during processing, or any error preventing the expected response from the
- * card.
- *
- * <p>It is also thrown when an automatic verification read, performed after a write operation on
- * storage cards that do not provide reliable write acknowledgment, fails to confirm that the
- * expected data was correctly stored.
+ * <p>This exception is typically thrown when the card responds with a status code that does not
+ * match the expected outcome of the executed command, suggesting that the operation could not be
+ * completed as intended (e.g., write or read command rejected by the card).
  *
  * @since 1.0.0
  */
-public final class CardIOException extends CardCommunicationException
+public final class SCInvalidCardResponseException extends InvalidCardResponseException
     implements StorageCardException {
 
   private final Integer blockAddress;
 
   /**
-   * Creates a new exception indicating a card communication error during the execution of a storage
+   * Creates a new exception indicating a card status code error during the execution of a storage
    * card command.
    *
    * @param blockAddress The block address involved in the error, or {@code null} if not relevant.
    * @param message The message describing the exception context.
    * @since 1.0.0
    */
-  public CardIOException(Integer blockAddress, String message) {
+  public SCInvalidCardResponseException(Integer blockAddress, String message) {
     super(message);
     this.blockAddress = blockAddress;
   }
 
   /**
-   * Creates a new exception indicating a card communication error during the execution of a storage
+   * Creates a new exception indicating a card status code error during the execution of a storage
    * card command, with an underlying cause.
    *
    * @param blockAddress The block address involved in the error, or {@code null} if not relevant.
@@ -54,7 +51,7 @@ public final class CardIOException extends CardCommunicationException
    * @param cause The underlying cause of the exception.
    * @since 1.0.0
    */
-  public CardIOException(Integer blockAddress, String message, Throwable cause) {
+  public SCInvalidCardResponseException(Integer blockAddress, String message, Throwable cause) {
     super(message, cause);
     this.blockAddress = blockAddress;
   }