+++ src/main/java/org/eclipse/keypop/storagecard/SCReaderCommunicationException.java
@@ -11,34 +11,46 @@
  ************************************************************************************** */
 package org.eclipse.keypop.storagecard;
 
-import org.eclipse.keypop.reader.selection.InvalidCardResponseException;
+import org.eclipse.keypop.reader.ReaderCommunicationException;
 import org.eclipse.keypop.storagecard.card.StorageCard;
 
 /**
- * Indicates that a command returned an unexpected or invalid status while interacting with the
- * {@link StorageCard}.
+ * Indicates an input/output error that occurred while communicating with the card reader.
  *
- * <p>This exception is typically thrown when the card responds with a status code that does not
- * match the expected outcome of the executed command, suggesting that the operation could not be
- * completed as intended (e.g., write or read command rejected by the card).
+ * <p>This exception reflects low-level reader communication failures, such as a lost connection,
+ * hardware malfunction, or driver-related issues, preventing the command from being transmitted to
+ * the {@link StorageCard}.
  *
  * @since 1.0.0
  */
-public final class UnexpectedCommandStatusException extends InvalidCardResponseException
+public final class SCReaderCommunicationException extends ReaderCommunicationException
     implements StorageCardException {
 
   private final Integer blockAddress;
 
   /**
-   * Creates a new exception indicating a card status code error during the execution of a storage
-   * card command, with an underlying cause.
+   * Creates a new exception indicating a reader communication error during the execution of a
+   * storage card command.
+   *
+   * @param blockAddress The block address involved in the error, or {@code null} if not relevant.
+   * @param message The message describing the exception context.
+   * @since 1.0.0
+   */
+  public SCReaderCommunicationException(Integer blockAddress, String message) {
+    super(message);
+    this.blockAddress = blockAddress;
+  }
+
+  /**
+   * Creates a new exception indicating a reader communication error during the execution of a
+   * storage card command, with an underlying cause.
    *
    * @param blockAddress The block address involved in the error, or {@code null} if not relevant.
    * @param message The message describing the exception context.
    * @param cause The underlying cause of the exception.
    * @since 1.0.0
    */
-  public UnexpectedCommandStatusException(Integer blockAddress, String message, Throwable cause) {
+  public SCReaderCommunicationException(Integer blockAddress, String message, Throwable cause) {
     super(message, cause);
     this.blockAddress = blockAddress;
   }